exlcudes request v2.77.0 or greater
this was a breaking change in a dependency, which unfortunately was
not treated as such with respect to semver.

in v2.77.0, request drops support for node versions 0.8.x - 4.x. this
package still supports node version 0.12.x, and therefore we need to
lock down the dependency version. as a cosequence, we will not
recieve bug fixes in that dependency, **this should motivate a
quicker deprecation of node 0.12**.