simplify scripts and update dev dependencies.
note that eslint was removed. tslint can lint javascript files now. getting
this to actually run requires specifiying rules again in the "jsRules" property of
`tslint.json`. this would be a waste of time right now since we'll be transitioning
all the source files to typescript. so for now, the most appropriate action is just
to remove eslint and its related config.

not everything was tested to work. `npm test` runs, but not until completion (suspect
that its because of dependencies that behave differenly). `npm run report-coverage`
launches the uploader but since there wasn't a complete report seems to bail out.
`npm run build:clean` has directories that we're not 100% sure would exist.
`npm run make-docs` doesn't even begin to work with the dependency that was removed,
we'll be looking into a better reference docs generation solution.